float32

+++ matcher.py
@@ -102,10 +102,8 @@ def compatibility(master, candidate, threshold=None):
     if threshold is None:
         threshold = 0.022
 
-    #master = np.array(master, dtype=np.float32)
-    #candidate = np.array(candidate, dtype=np.float32)
-    master = np.array(master)
-    candidate = np.array(candidate)
+    master = np.array(master, dtype=np.float32)
+    candidate = np.array(candidate, dtype=np.float32)
 
     # fft
     frequency = int(np.ceil(1 / threshold))
